Create Linked_List_Delete_Nth_Node
This program adds nodes (which are like little boxes) to the linked list, then deletes a node from the list. The goal is to understand how data flows and is managed in memory (like a well-organized shelf!).

Note, that we create a sample list, and we count from 20 to-->11, because we add each new node at the beginning.

When deleting a node, we choose the node by its position from the beginning of the list. For example, if you want to delete the 1st node, the 2nd node, the 3rd node, or the Nth node, you specify its position, and the program will remove it accordinglyon.

You can add, remove, and explore the list—it's like playing Tetris with data! 🧑‍💻 This was a fun project where I learned about pointers, memory management, and the importance of cleaning up after myself (no one likes a messy list, right?).

Feel free to play around with the code, make changes, and let me know if you have any questions or suggestions