Create Linked_List_Find_Position_of_node_with_max_value
Explanation:
Linked List Creation: We are creating a linked list by adding nodes one by one with values from 8 down to 1. Each new node is inserted at the beginning of the list.
Finding the Largest Value: The function bigger goes through the list and finds the largest value and its position (1st, 2nd, etc.).
Memory Cleanup: The delete function frees the memory by iterating over the list and freeing each node.
How the list looks:
Initially, the list will be in the order 8 -> 7 -> 6 -> ... -> 1 (since we insert at the beginning).
After modifying the second node’s value to 15, the list will look like this: 8 -> 15 -> 6 -> 5 -> ... -> 1.
Key Notes:
bigger(): Finds the position of the largest value in the list.
delete(): Frees the memory of all the nodes in the list.
The program prints the list and shows the position of the largest value.

You can modify each element you want to test it. You can also not modify anything, it will telle you that the first node has the greatest value which is true. (8)